conda r-base

+++ files/setup-radiant.sh
@@ -16,25 +16,25 @@ fi
 
 NCPUS=${NCPUS:--1}
 
-export DEBIAN_FRONTEND=noninteractive
-apt-get update -qq && apt-get -y --no-install-recommends install \
-    libicu-dev \
-    pandoc \
-    zlib1g-dev \
-    libglpk-dev \
-    libgmp3-dev \
-    libxml2-dev \
-    cmake \
-    git \
-    libharfbuzz-dev \
-    libfribidi-dev \
-    libfreetype6-dev \
-    libpng-dev \
-    libtiff5-dev \
-    libjpeg-dev \
-    libsnappy-dev \
-    libre2-dev \
-    && rm -rf /var/lib/apt/lists/*
+if [ "$(which R)" != "/opt/conda/bin/R" ]; then
+  export DEBIAN_FRONTEND=noninteractive
+  apt-get update -qq && apt-get -y --no-install-recommends install \
+      libicu-dev \
+      zlib1g-dev \
+      libglpk-dev \
+      libgmp3-dev \
+      libxml2-dev \
+      cmake \
+      git \
+      libharfbuzz-dev \
+      libfribidi-dev \
+      libfreetype6-dev \
+      libpng-dev \
+      libtiff5-dev \
+      libjpeg-dev \
+      libcurl4-openssl-dev \
+      && rm -rf /var/lib/apt/lists/*
+fi
 
 R -e "install.packages('igraph', repo='${CRAN}', Ncpus=${NCPUS})" \
   -e "install.packages(c('radiant', 'png', 'bslib', 'gitgadget', 'miniUI', 'webshot', 'tinytex', 'svglite'), repo='${CRAN}', Ncpus=${NCPUS})" \
@@ -52,19 +52,7 @@ R -e "install.packages('igraph', repo='${CRAN}', Ncpus=${NCPUS})" \
   -e "remotes::install_github('radiant-rstats/radiant.model', upgrade = 'never')" \
   -e "remotes::install_github('radiant-rstats/radiant.multivariate', upgrade = 'never')" \
   -e "remotes::install_github('radiant-rstats/radiant', upgrade = 'never')" \
-  -e "remotes::install_github('radiant-rstats/radiant.update', upgrade = 'never')"
-
-# arrow install from source is not currently working on aarch64
-# https://issues.apache.org/jira/projects/ARROW/issues/ARROW-17374?filter=allopenissues
-
-if [ "$(uname -m)" != "aarch64" ]; then
-  R -e "install.packages(c('duckdb', 'arrow'), repo='${CRAN}', Ncpus=${NCPUS})"
-else
-  # based on https://github.com/duckdb/duckdb/issues/3049#issuecomment-1096671708
-  cp -a /usr/local/lib/R/etc/Makeconf /usr/local/lib/R/etc/Makeconf.bak;
-  sed -i 's/fpic/fPIC/g' /usr/local/lib/R/etc/Makeconf;
-  R -e "options(HTTPUserAgent = sprintf('R/%s R (%s)', getRversion(), paste(getRversion(), R.version['platform'], R.version['arch'], R.version['os']))); Sys.setenv('ARROW_R_DEV' = TRUE); install.packages(c('duckdb', 'arrow'), repo='${CRAN}', Ncpus=${NCPUS})"
-  mv /usr/local/lib/R/etc/Makeconf.bak /usr/local/lib/R/etc/Makeconf;
-fi
+  -e "remotes::install_github('radiant-rstats/radiant.update', upgrade = 'never')" \
+  -e "Sys.setenv('ARROW_PARQUET'='ON'); install.packages(c('duckdb', 'arrow'), repo='${CRAN}', Ncpus=${NCPUS})" \
 
 rm -rf /tmp/downloaded_packages
